ex 4.2 end
If we define the median of a sequence as “a number so that exactly as
many elements come before it in the sequence as come after it,” fix the
program in §4.6.3 so that it always prints out a median. Hint: A median
need not be an element of the sequence.